engine/pyproject.toml
[project]
name = "zaseon-engine"
version = "0.1.0"
description = "ZASEON AI Code Scanner & Smart Contract Auditor Engine"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "pygithub>=2.5.0",
    "gitpython>=3.1.43",
    "tree-sitter>=0.23.0",
    "anthropic>=0.40.0",
    "openai>=1.57.0",
    "pgvector>=0.3.6",
    "boto3>=1.35.0",
    "docker>=7.1.0",
    "weasyprint>=63.0",
    "jinja2>=3.1.4",
    "web3>=7.6.0",
    "py-solc-x>=2.0.0",
    "slither-analyzer>=0.10.0",
    "pyjwt>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    "email-validator>=2.1.0",
    "flower>=2.0.0",
    "cryptography>=42.0.0",
    "prometheus-client>=0.21.0",
    "scikit-learn>=1.5.0",
    "z3-solver>=4.12.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
zaseon = "engine.cli.main:main"

[project.entry-points."zaseon.cli"]
serve = "engine.api.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
    "bandit>=1.7.0",
    "safety>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101"]
